worker/requirements.txt
psycopg2-binary>=2.9
python-dotenv>=1.0
opencv-python-headless>=4.8
openai>=1.0
